+++ src/app/layout.tsx
@@ -1,13 +1,21 @@
+'use client';
+import { ReactNode, useState } from 'react';
+
 import localFont from 'next/font/local';
 
-import type { Metadata } from 'next';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+// import type { Metadata } from 'next';
 
 import Header from '@/components/header';
+
+// Styles
+import 'maplibre-gl/dist/maplibre-gl.css';
 import '@/styles/globals.css';
-export const metadata: Metadata = {
-  title: 'Hub - Open-Earth-Monitor Cyberinfrastructure',
-  description: '...',
-};
+
+// export const metadata: Metadata = {
+//   title: 'Hub - Open-Earth-Monitor Cyberinfrastructure',
+//   description: '...',
+// };
 
 const satoshi = localFont({
   src: [
@@ -24,13 +32,17 @@ const satoshi = localFont({
   ],
 });
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function RootLayout({ children }: { children: ReactNode }) {
+  const [queryClient] = useState(() => new QueryClient());
+
   return (
-    <html lang="en" className={`${satoshi.className} h-screen w-screen`}>
-      <body className="inter mx-auto h-screen max-w-7xl bg-brand-600">
-        <Header />
-        <main>{children}</main>
-      </body>
-    </html>
+    <QueryClientProvider client={queryClient}>
+      <html lang="en" className={`${satoshi.className} h-screen w-screen`}>
+        <body className="inter mx-auto h-screen max-w-7xl bg-brand-600">
+          <Header />
+          <main>{children}</main>
+        </body>
+      </html>
+    </QueryClientProvider>
   );
 }