+++ src/app/map/page.tsx
@@ -1,14 +1,18 @@
 import type { Metadata } from 'next';
 
+import Sidebar from '@/containers/sidebar';
+
+import Map from '@/doc-containers/map-implementation';
 export const metadata: Metadata = {
   title: 'Map - Open-Earth-Monitor Cyberinfrastructure',
   description: '...',
 };
 
-const Map: React.FC = () => (
-  <div>
-    <h1>Map</h1>
+const MapPage: React.FC = () => (
+  <div className="flex h-screen w-screen">
+    <Sidebar />
+    <Map />
   </div>
 );
 
-export default Map;
+export default MapPage;