+++ src/doc-containers/map-implementation/layer-manager/index.tsx
@@ -17,6 +17,8 @@ const LayerManagerContainer = ({ layers }: { layers: string[] }) => {
     <>
       {LAYERS_FILTERED?.map((layer, i) => {
         const LayerComponent: FC<LayerComponentProps> = LAYER_COMPONENTS[layer];
+        // We need to define where do we want to put the layer
+        // We want to put it before the custom-layers transparent backgrond
         const beforeId = i === 0 ? 'custom-layers' : `${layers[i - 1]}-layer`;
         return <LayerComponent key={layer} beforeId={beforeId} />;
       })}